deeplom2

+++ client/src/store/UserStore.js
@@ -1,4 +1,4 @@
-import {makeAutoObservable } from "mobx";
+import { makeAutoObservable } from "mobx";
 
 export default class UserStore {
     constructor() {
@@ -13,7 +13,6 @@ export default class UserStore {
     }
     setUser(user) {
         this._user = user
-        console.log(user)
     }
 
     setCart(cart) {
@@ -21,28 +20,48 @@ export default class UserStore {
     }
 
     addCartItem(cartItemToAdd) {
-         const existingCartItem = this._cart.find(
+        const existingCartItem = this._cart.find(
             cartItem => cartItem.id === cartItemToAdd.id
         );
 
         if (existingCartItem) {
-            this._cart = this._cart.map(cartItem => 
-            cartItem.id === cartItemToAdd.id
-                ? { ...cartItem, quantity: cartItem.quantity + 1}
-                : cartItem
+            this._cart = this._cart.map(cartItem =>
+                cartItem.id === cartItemToAdd.id
+                    ? { ...cartItem, quantity: cartItem.quantity + 1 }
+                    : cartItem
             );
             localStorage.setItem('cart', JSON.stringify(this._cart))
             return
         }
 
-        this._cart =  [...this._cart, { ...cartItemToAdd, quantity: 1 }];
+        this._cart = [...this._cart, { ...cartItemToAdd, quantity: 1 }];
         localStorage.setItem('cart', JSON.stringify(this._cart))
+    }
+    deleteCartItem(cartItemToDelite) {
+        const existingCartItem = this._cart.find(
+            cartItem => cartItem.id === cartItemToDelite.id
+        );
+        if (existingCartItem.quantity === 1){
+            this._cart = this._cart.filter(item => item.id !== cartItemToDelite.id)
+            localStorage.setItem('cart', JSON.stringify(this._cart))
+            return
+        }
+        
+        if (existingCartItem) {
+            this._cart = this._cart.map(cartItem =>
+                cartItem.id === cartItemToDelite.id
+                    ? { ...cartItem, quantity: cartItem.quantity - 1 }
+                    : cartItem
+            );
+            localStorage.setItem('cart', JSON.stringify(this._cart))
+            return
+        }
+
+       
     }
 
-    removeCartItem(id) {
-        this._cart = this._cart.filter(item => item.id !== id)
-        localStorage.setItem('cart', JSON.stringify(this._cart))
-    }
+
+
 
     get cart() {
         return this._cart
